# Conflicts:
#	2Dshooter/src/game/gameView/GameView.java

+++ 2Dshooter/src/game/core/PowerUp.java
@@ -9,7 +9,7 @@ public interface PowerUp {
 
     void setEffect(Player player);
 
-    int getX();
-    int getZ();
+    float getX();
+    float getZ();
 
 }
